decode noise is now a single byte ratio of signal delta to max error margin

+++ src/Quest_IR_Receiver.h
@@ -27,7 +27,7 @@ public:
   IRDecodeState decodeState;
   uint8_t decodedBits[QIR_BUFFER_SIZE];
   uint16_t decodedBitCount;
-  uint16_t decodeMaxNoise;
+  uint8_t decodeMaxNoise;
 
   void begin(uint8_t irPin);
   void enable();
